Address PR #887 review: shared portal-projects types/fetch, page layout, membership copy
Resolve the confirmed adversarial review findings on the Projects section (#816):

- Rule 9 (type placement): move Project / GetMyProjectsResponse /
  GetMyProjectsErrorCode into src/shared/types/portal.ts and consume via
  @shared/types/portal from both layers. backend/types and ui/types now
  re-export them instead of duplicating the declarations, and ProjectsPage
  imports from @shared/types/portal rather than reaching through to
  ../../../backend/types.

- Rule 7 / Rule 10 (DRY): extract the /projects/summaries fetch + the
  ProjectSummaryDto contract into src/backend/services/portal/portal-projects.ts
  (fetchProjectSummaries + mapProjectsResponse). Both the new get-my-projects
  handler and PromptManager.getRemotePrompts now call the shared fetcher, so the
  endpoint contract has a single owner and can't drift. Removes the duplicated
  https.request block from both call sites.

- Page-layout convention: wrap ProjectsPage in the shared feature-page container
  (<main className="z-10 flex flex-col p-8 h-full gap-6 w-full min-w-0">) so the
  route matches HomePage / WorkflowPage instead of rendering flush to the edge.

- AC2 membership scoping: /projects/summaries is documented by #691 as returning
  ALL projects and the user-scoped backend endpoint (SSWConsulting/SSW.YakShaver#3775)
  is still open, so soften the desktop copy from membership claims ("projects
  you're a member of", "You're not a member of any projects yet") to neutral
  wording until scoping is confirmed.
