+++ src/backend/services/prompt/prompt-manager.ts
@@ -2,6 +2,7 @@ import https from "node:https";
 import { config } from "../../config/env";
 import { formatErrorMessage } from "../../utils/error-utils";
 import { IdentityServerAuthService } from "../auth/identity-server-auth";
+import { fetchProjectSummaries, type ProjectSummaryDto } from "../portal/portal-projects";
 import { CustomPromptStorage } from "../storage/custom-prompt-storage";
 
 // Define the Prompt interface that consolidates local and remote prompts
@@ -10,13 +11,6 @@ export interface PromptSummary {
   name: string;
   description?: string;
   source: "local" | "remote";
-}
-
-// Define the ProjectSummaryDto interface based on the API response
-interface ProjectSummaryDto {
-  id: string;
-  title: string;
-  description?: string;
 }
 
 // Define the ProjectDto interface based on the C# class
@@ -105,55 +99,12 @@ export class PromptManager {
         return [];
       }
 
-      const apiUrl = config.portalApiUrl();
-      const url = new URL(apiUrl);
-      const hostname = url.hostname;
-      const port = url.port ? Number.parseInt(url.port, 10) : url.protocol === "https:" ? 443 : 80;
-      // Append /projects/summaries to the base API path
-      const path = `${url.pathname.replace(/\/$/, "")}/projects/summaries`;
-
-      const data = await new Promise<ProjectSummaryDto[]>((resolve, reject) => {
-        const options = {
-          hostname: hostname,
-          port: port,
-          path: path,
-          method: "GET",
-          headers: {
-            Authorization: `Bearer ${accessToken}`,
-            "Content-Type": "application/json",
-          },
-        };
-
-        const req = https.request(options, (res) => {
-          let responseData = "";
-
-          res.on("data", (chunk) => {
-            responseData += chunk;
-          });
-
-          res.on("end", () => {
-            if (res.statusCode && res.statusCode >= 200 && res.statusCode < 300) {
-              try {
-                const parsedData = JSON.parse(responseData);
-                // The API might return { data: [...] } or just [...]
-                // Adjust based on actual API response structure.
-                // Assuming it returns an array of project summaries which act as prompts.
-                resolve(Array.isArray(parsedData) ? parsedData : []);
-              } catch (error) {
-                reject(new Error(`Failed to parse JSON response: ${formatErrorMessage(error)}`));
-              }
-            } else {
-              reject(new Error(`API call failed: ${res.statusCode} ${res.statusMessage}`));
-            }
-          });
-        });
-
-        req.on("error", (error) => {
-          reject(error);
-        });
-
-        req.end();
-      });
+      // Shared portal-projects fetch — single owner of the /projects/summaries contract.
+      const parsed = await fetchProjectSummaries(accessToken);
+      // This consumer treats an unexpected body as "no remote prompts" rather than an error.
+      const data: ProjectSummaryDto[] = Array.isArray(parsed)
+        ? (parsed as ProjectSummaryDto[])
+        : [];
 
       // Map remote data to Prompt interface
       return data.map((item: ProjectSummaryDto) => ({
