+++ src/ui/src/components/projects/NoProjects.tsx
@@ -4,10 +4,9 @@ export function NoProjects() {
   return (
     <Empty>
       <EmptyHeader className="gap-4 items-start!">
-        <EmptyTitle>📁 - You're not a member of any projects yet</EmptyTitle>
+        <EmptyTitle>📁 - No projects to show yet</EmptyTitle>
         <EmptyDescription>
-          Projects you're added to will appear here. Ask a project owner to add you, or check back
-          later.
+          Projects will appear here once they're available. Check back later.
         </EmptyDescription>
       </EmptyHeader>
     </Empty>
